+++ app/src/main/AndroidManifest.xml
@@ -22,7 +22,7 @@
         android:theme="@style/Theme.Score"
         tools:targetApi="31">
         <activity
-            android:name=".onBoarding.OnboardingActivity"
+            android:name=".OnBoarding.OnboardingActivity"
             android:exported="false" />
         <activity
             android:name=".MainActivity"
