+++ app/src/main/java/com/project/score/OnBoarding/Onboarding1Fragment.kt
@@ -1,11 +1,10 @@
-package com.project.score.onBoarding
+package com.project.score.OnBoarding
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import com.project.score.R
 import com.project.score.databinding.FragmentOnboarding1Binding
 
 class Onboarding1Fragment : Fragment() {
