+++ app/src/main/java/com/project/score/OnBoarding/Onboarding2Fragment.kt
@@ -1,23 +1,22 @@
-package com.project.score.onBoarding
+package com.project.score.OnBoarding
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import com.project.score.R
-import com.project.score.databinding.FragmentOnboarding4Binding
+import com.project.score.databinding.FragmentOnboarding2Binding
 
-class Onboarding4Fragment : Fragment() {
+class Onboarding2Fragment : Fragment() {
 
-    lateinit var binding: FragmentOnboarding4Binding
+    lateinit var binding: FragmentOnboarding2Binding
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
 
-        binding = FragmentOnboarding4Binding.inflate(layoutInflater)
+        binding = FragmentOnboarding2Binding.inflate(layoutInflater)
 
         return binding.root
     }
