+++ app/src/main/java/com/project/score/OnBoarding/Onboarding4Fragment.kt
@@ -1,22 +1,22 @@
-package com.project.score.onBoarding
+package com.project.score.OnBoarding
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import com.project.score.databinding.FragmentOnboarding2Binding
+import com.project.score.databinding.FragmentOnboarding4Binding
 
-class Onboarding2Fragment : Fragment() {
+class Onboarding4Fragment : Fragment() {
 
-    lateinit var binding: FragmentOnboarding2Binding
+    lateinit var binding: FragmentOnboarding4Binding
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
 
-        binding = FragmentOnboarding2Binding.inflate(layoutInflater)
+        binding = FragmentOnboarding4Binding.inflate(layoutInflater)
 
         return binding.root
     }
