+++ app/src/main/res/layout/activity_onboarding.xml
@@ -5,7 +5,7 @@
     android:id="@+id/main"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".onBoarding.OnboardingActivity">
+    tools:context=".OnBoarding.OnboardingActivity">
 
     <Button
         android:id="@+id/button"
