+++ app/src/main/res/layout/fragment_onboarding.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context=".onBoarding.OnboardingFragment">
+    tools:context=".OnBoarding.OnboardingFragment">
 
     <Button
         android:id="@+id/button_next"
@@ -43,13 +43,14 @@
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toBottomOf="@+id/textView_skip" />
 
-    <com.tbuonomo.viewpagerdotsindicator.DotsIndicator
+    <com.tbuonomo.viewpagerdotsindicator.WormDotsIndicator
         android:id="@+id/dots_indicator"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:layout_gravity="center_horizontal"
         android:layout_marginBottom="20dp"
-        app:dotsColor="@color/grey3"
+        app:dotsColor="@color/main"
+        app:dotsStrokeColor="@color/grey3"
         app:dotsCornerRadius="8dp"
         app:dotsSize="8dp"
         app:dotsSpacing="16dp"
@@ -59,6 +60,6 @@
         app:layout_constraintStart_toStartOf="parent"
         app:selectedDotColor="@color/main">
 
-    </com.tbuonomo.viewpagerdotsindicator.DotsIndicator>
+    </com.tbuonomo.viewpagerdotsindicator.WormDotsIndicator>
 
 </androidx.constraintlayout.widget.ConstraintLayout>
