+++ app/src/main/res/layout/fragment_onboarding2.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context=".onBoarding.Onboarding2Fragment">
+    tools:context=".OnBoarding.Onboarding2Fragment">
 
     <TextView
         android:id="@+id/textView_tag"
