+++ app/src/main/res/layout/fragment_onboarding3.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context=".onBoarding.Onboarding3Fragment">
+    tools:context=".OnBoarding.Onboarding3Fragment">
 
     <TextView
         android:id="@+id/textView_tag"
